+++ laravel/public/qiantai/css/style.css
@@ -3579,3 +3579,45 @@ ul.feature_list.feature_list1 {
     .table th.text-l,.table td.text-l{text-align:left!important}/*单独列居左*/
     .table th.text-c,.table td.text-c{text-align:center!important}/*单独列居中*/
     .table th.text-r,.table td.text-r{text-align:right!important}/*单独列居右*/
+
+
+
+
+.black_overlay{  
+ display: none;  
+ position: absolute;  
+ top: 0%;  
+ left: 0%;  
+ width: 100%;  
+ height: 100%;  
+ background-color: black;  
+ z-index:1001;  
+ -moz-opacity: 0.8;  
+ opacity:.100;  
+ filter: alpha(opacity=80);
+}  
+.white_content { 
+ display: none;  
+ position: absolute;  
+ top: 10%;  
+ left: 30%;  
+ width: 40%;  
+ height: 65%;  
+ border: 5px solid lightblue;  
+ background-color: white;
+ z-index:1002;  
+ overflow: auto;  
+}  
+.white_content_small {  
+ display: none;  
+ position: absolute;  
+ top: 20%;  
+ left: 30%;  
+ width: 40%;  
+ height: 50%;  
+ border: 16px solid lightblue;  
+ background-color: white;  
+ z-index:1002;
+ overflow: auto;  
+}  
+
