+++ laravel/resources/views/index/personal_data.blade.php
@@ -8,6 +8,12 @@
 <!-- 内容输入区 -->
 @section('content')
 <style type="text/css">
+      .star-bar{font-size:0; line-height:0}
+      .star-bar .star{display:inline-block;text-align:center}
+      /*中*/
+      .size-M img{ width:24px;height:24px}
+      /*小*/
+      .size-S img{width:16px;height:16px}
       #login
       {
           display:none;
@@ -39,6 +45,7 @@
 <link rel="stylesheet" type="text/css" href="/qiantai/css/h-ui/iconfont.css" />
 <link rel="stylesheet" type="text/css" href="/qiantai/css/h-ui/skin.css" id="skin" />
 <link rel="stylesheet" type="text/css" href="/qiantai/css/h-ui/style.css" />
+
 
 
 <div class="page-container">
@@ -104,17 +111,62 @@
 
 
             <div class="tabCon">
-                这是订单
+                <div class="mt-20">
+                         <table class="table table-border table-bg">
+                            <thead class="text-c">
+                                <tr>
+                                  <th>订单号</th>
+                                  <th>酒店名称</th>
+                                  <th>房间号</th>
+                                  <th>房间类型</th>
+                                  <th>总价格</th> 
+                                  <th>支付类型</th> 
+                                  <th>支付状态</th>
+                                  <th>入住时间</th>
+                                  <th>离开时间</th> 
+                                  <th>客户联系方式</th>
+                                  <th></th>
+                                </tr>
+                            </thead>
+                            <tbody id="tbody">
+                            @foreach($order as $v)
+                                <tr class="text-c" data-id="{{$v->hotel_id}}">
+                             
+                                  <td>{{$v->order_sn}}</td>
+                                  @foreach($hotel_name as $val)
+                                  <td>{{$val->hotel_name}}</td>
+                                  @endforeach
+
+                                  @foreach($room_num as $ve)
+                                  <td>{{$ve->room_number}}</td>
+                                  @endforeach
+
+                                  @foreach($room_type as $va)
+                                  <td>{{$va->room_type_name}}</td>
+                                  @endforeach
+                                  <td>{{$v->total_price}}</td>
+                                  <td>@if($v->pay_type==1)微信支付@elseif($v->pay_type==2)支付宝支付@elseif($v->pay_type==3)百度钱包支付@($v->pay_type==4)余额支付@endif</td>
+                                  <td>@if($v->pay_status==0)未支付@elseif($v->pay_status==1)已支付@endif</td>
+                                  <td>{{$v->check_time}}</td>
+                                  <td>{{$v->end_time}}</td>
+                                  <td>{{$v->cell_phone}}</td>
+                                  <td><input id="Button1" type="button" class="btn btn-primary radius" value="评价" onclick="ShowDiv('MyDiv','fade')" ></td>
+                                </tr>
+                            @endforeach
+                              </tbody>
+                        </table>
+                </div>     
             </div>
             <div class="tabCon">
-                这是积分
+                   
             </div>
             <div class="tabCon">
                 这是余额
             </div>
             <div class="tabCon">
                 <div class="cl pd-5 bg-1 bk-gray mt-20">
-                <a href="javascript:show()">使用规则 </a> <span class="r">共有兑换券：<strong><?php echo count($detailed); ?></strong> 张</span> </div>
+                <input onclick="javascript:show()" type="button" value="使用规则" class="btn btn-primary-outline radius">
+                <span class="r">共有兑换券：<strong><?php echo count($detailed); ?></strong> 张</span> </div>
                 <div class="mt-20">
                          <table class="table table-border table-bg">
                             <thead class="text-c">
@@ -152,6 +204,7 @@
         </div>        
     </form>
 </div>
+
 <!-- 弹出层 -->
 <div id="login">
 <a href="javascript:hide()">✖</a>
@@ -171,7 +224,45 @@
      </div>
   <div id="over"></div>
 <!-- 弹出层结束 -->
+             
+<style>
+  #XX{
+    text-align:center;
+    margin-right:1px;
+    display: inline-block;
+    width:30px;
+    height: 30px;
+    line-height: 30px;
+  }
+  #XX:hover{background-color: #EE0000}
+</style>
+<!-----评价弹出层---->
+<div id="fade" class="black_overlay">  
+</div>  
+ <div id="MyDiv" class="white_content">  
+  <div style="text-align: right; cursor: default; height: 40px;">  
+   <span id="XX"  style="font-size: 16px; cursor:pointer;" class="close_div" onclick="CloseDiv('MyDiv','fade')">✖</span>  
+  </div>  
+ 
+      <div style="width:500px; height:50px;">
+        <span style='margin-left:13px' class="f-l f-15 va-m">描述相符：</span>
+        <div id="star-1" class="star-bar size-M f-l mr-10 va-m" style="width:250px;"></div>
+        <strong id="result-1" class="f-l va-m"></strong>
+      </div>
 
+      <div style="width:100%">
+        <span style='margin-left:13px' class="f-l f-15 va-m">评价内容：</span>
+        <input type="hidden" value="" class="hotel_id">
+        <textarea id="assess_desc" cols="60" rows="10"></textarea>
+      </div>
+
+      <div style="text-align:center;margin-top:20px;">
+        <input type="button" value="评价" class="btn btn-primary radius assess">&nbsp;&nbsp;&nbsp;&nbsp;
+        <input class="btn btn-default radius" type="button" value="取消">
+      </div>
+
+ </div>  
+<!-----评价弹结束---->
 <!--_footer 作为公共模版分离出去-->
 <script type="text/javascript" src="/qiantai/js/h-ui/jquery.min.js"></script>
 <script type="text/javascript" src="/qiantai/js/h-ui/layer.js"></script>
@@ -185,63 +276,126 @@
 <script type="text/javascript" src="/qiantai/js/h-ui/messages_zh.js"></script>
 <!-- Resource jQuery -->
 <script type="text/javascript">
-var login = document.getElementById('login');
-var over = document.getElementById('over');
-    function show()
-   {
-      login.style.display = "block";
-       over.style.display = "block";
-    }
-    function hide()
-    {
-         login.style.display = "none";
-         over.style.display = "none";
-    }
- 
+    star_num="";
+    //弹出隐藏层  
+    function ShowDiv(show_div,bg_div){  
+     document.getElementById(show_div).style.display='block';  
+     document.getElementById(bg_div).style.display='block' ;  
+     var bgdiv = document.getElementById(bg_div);  
+     bgdiv.style.width = document.body.scrollWidth;   
+     $("#"+bg_div).height($(document).height());
 
-$(function(){
-    $('.skin-minimal input').iCheck({
-        checkboxClass: 'icheckbox-blue',
-        radioClass: 'iradio-blue',
-        increaseArea: '20%'
+    };  
+    //关闭弹出层  
+    function CloseDiv(show_div,bg_div)  
+    {  
+     document.getElementById(show_div).style.display='none';  
+     document.getElementById(bg_div).style.display='none';  
+    };  
+
+   $("#tbody").delegate("#Button1","click",function(){
+       var hotel_id  = $(this).parents('tr').data('id');
+       $(".hotel_id").val(hotel_id);
+   })
+
+   $(".close_div").click(function(){
+      CloseDiv('MyDiv','fade');
+   });
+ 
+   $('.assess').click(function(){
+        var assess_desc = $("#assess_desc").val();
+        var hotel_id=$(".hotel_id").val();
+        $.ajax({
+            url:'assess_desc',
+            data:{'assess_desc':assess_desc,'hotel_id':hotel_id,'assess_num':star_num},
+            type:'post',
+            headers: {'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')},
+            sussess:function(msg){
+                if(msg==1){
+                  CloseDiv('MyDiv','fade');
+                }
+            }
+        })
     });
-    $("#tab-system").Huitab({
-        index:0
+
+
+
+    var login = document.getElementById('login');
+    var over = document.getElementById('over');
+        function show()
+       {
+          login.style.display = "block";
+           over.style.display = "block";
+        }
+        function hide()
+        {
+             login.style.display = "none";
+             over.style.display = "none";
+        }
+     
+    //兑换券弹出层操作
+    $(function(){
+        $('.skin-minimal input').iCheck({
+            checkboxClass: 'icheckbox-blue',
+            radioClass: 'iradio-blue',
+            increaseArea: '20%'
+        });
+        $("#tab-system").Huitab({
+            index:0
+        });
     });
-});
 
     //图片处理
-function uploadPic(){
-    var pic = $('#upload_file')[0].files[0];
-    var fd = new FormData();
-    fd.append('uploadFile', pic);
-    var token=$("#token").val();
-    var obj=$(this);
-    $.ajax({
-        headers: {'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')},
-        url:"user_img",
-        type:"post",
-        'X-XSRF-TOKEN':"token",
-        // Form数据
-        data: fd,
-        cache: false,
-        contentType: false,
-        processData: false,
-        success:function(msg){
-           $('.img').html("<img src="+msg+" alt=''width='100px' height='97px'>");
-           $("input[name='img']").val(msg);
-        }
-    });
-                        
-}
+    function uploadPic(){
+        var pic = $('#upload_file')[0].files[0];
+        var fd = new FormData();
+        fd.append('uploadFile', pic);
+        var token=$("#token").val();
+        var obj=$(this);
+        $.ajax({
+            headers: {'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')},
+            url:"user_img",
+            type:"post",
+            'X-XSRF-TOKEN':"token",
+            // Form数据
+            data: fd,
+            cache: false,
+            contentType: false,
+            processData: false,
+            success:function(msg){
+               $('.img').html("<img src="+msg+" alt=''width='100px' height='97px'>");
+               $("input[name='img']").val(msg);
+            }
+        });
+                          
+    }
     $(".edit").click(function(){
        $("input").removeAttr("readonly",'false');
     });
+    
+    //星星评价
+    $(function(){
+      $("#star-1").raty({
+        hints: ['1','2', '3', '4', '5'],//自定义分数
+        starOff: 'iconpic-star-S-default.png',//默认灰色星星
+        starOn: 'iconpic-star-S.png',//黄色星星
+        path: '/qiantai/images/star',//可以是相对路径
+        number: 5,//星星数量，要和hints数组对应
+        showHalf: true,
+        targetKeep : true,
+        click: function (score, evt) {
+          star_num=score;
+          $("#result-1").html('你的评分是'+score+'分');
+        }
+      });
+    });
 
+    
+
+
+   
 
  </script>
-
-
 @endsection
 
 @section('footer')
